# County hovers + no circle sort = better category switch performance