--- a/js/build_map.js
+++ b/js/build_map.js
@@ -102,6 +102,10 @@ function create_map() {
   
   // add the main, active map features
   addStates(map, stateData);
+  
+  // add placeholder group for county boundaries
+  map.append('g').classed("county-bounds", true);
+  
   if(activeCategory === 'piechart') {
     // prepare and then use the centroid data for presentation as pie charts
     pieFormData = pieData(countyCentroids);
@@ -112,12 +116,9 @@ function create_map() {
     pieFormData = pieData(countyCentroids);
   }
   
-  // get started downloading county data right away.
-  // for now, pretend that we know that state '01' is the most likely state
-  // for the user to click on; we could make this dynamic in the future.
-  loadCountyData("AL", function(error, data) {
-    if (error) throw error;
-  });
+  // load all county data - it's OK if it's not done right away
+  // it should be loaded by the time anyone tries to hover!
+  showCounties('USA');
 
 }