--- a/js/counties.js
+++ b/js/counties.js
@@ -92,9 +92,13 @@ function displayCountyData(error, activeCountyData) {
       .attr('d', buildPath)
       .on("mouseover", function(d) {
         highlightCounty(this); 
+        // OK to use global var activeCategory which only changes on click 
+        // because people won't be able to hover on tooltips at the same time as hovering buttons
+        showToolTip(this, d, activeCategory); 
       })
       .on("mouseout", function(d) { 
         unhighlightCounty(this); 
+        hideToolTip(this, d, activeCategory);
       });
     
     // update