--- a/js/counties.js
+++ b/js/counties.js
@@ -92,6 +92,7 @@ function displayCountyData(error, activeCountyData) {
       .style("stroke-width", 0.2)
       .attr('d', buildPath)
       .on("mouseover", function(d) {
+        highlightState(d3.select("#"+d.properties.STATE_ABBV));
         highlightCounty(this); 
         highlightCircle(d3.select("#"+"circle-"+d.properties.GEOID));
         showToolTip(d, activeCategory); 
@@ -99,6 +100,7 @@ function displayCountyData(error, activeCountyData) {
         // because people won't be able to hover on tooltips at the same time as hovering buttons
       })
       .on("mouseout", function(d) { 
+        unhighlightState(d3.select("#"+d.properties.STATE_ABBV));
         unhighlightCounty(this);
         unhighlightCircle();
         hideToolTip();