--- a/js/map_utils.js
+++ b/js/map_utils.js
@@ -151,6 +151,20 @@ function unhighlightState() {
     .style('stroke-width', function(d) { return formatState('stroke-width', d, false); });
 }
 
+// on mouseover
+function highlightCounty(selection) {
+  console.log("highlighting");
+  d3.select(selection)
+    .style('fill', function(d) { return "darkgrey"; });
+}
+
+// on mouseout
+function unhighlightCounty(selection) {
+  console.log("unhighlighting");
+  d3.select(selection)
+    .style("fill", function(d) { return "transparent" });
+}
+
 // on click
 function zoomToFromState(data) {