--- a/layout/css/map.css
+++ b/layout/css/map.css
@@ -24,6 +24,11 @@
   stroke-width: 0;
 }
 
+.county-point,
+.pieslice {
+  pointer-events: none;
+}
+
 #slideraxis path,
 #slideraxis line {
   stroke: #fff;