Now all mouseovers and clicks are based on the county polygons, which are loaded from the start. Circles highlight when the county is hovered over, but they actually don't have pointer events on. Colors for states, counties, and highlights need work. You can also see all county boundaries when zoomed in. It currently displays all of them, but we can work on trying a clip to show which counties are in view and which aren't. This was already a lot of change and I didn't want to go too crazy. I think having all the counties on zoom helps with the cutoff circles a little bit. 

![countyhovers](https://user-images.githubusercontent.com/13220910/38140765-d2c32018-33fa-11e8-807d-a3910b316b90.gif)

Category switching is much faster now that we turned off sorting. Unfortunately, the full load time for the page has increased because all counties are loading at the national view. The time to be able to see circles on the map is slightly quicker than before (because no sorting is needed), but there's just a slight lag before you can start hovering on counties.

Performance results:

## Page load times

### BEFORE: Load times to `view=USA&category=total` (milliseconds):

total | scripts (all the yellow) | rendering (all the purple) | painting (all the green) | 'load' handler took __ms
-- | -- | -- | -- | --
1990 | 1178 | 44 | 8 | --
1991 | 1193 | 44 | 8 | --
1173 | 970 | 42 | 5 | --
2039 | 1231 | 45 | 9 | --
1196 | 994 | 37 | 5 | --

### AFTER: Load times to finish loading everything (so the county polygons)

total | scripts (all the yellow) | rendering (all the purple) | painting (all the green) | 'load' handler took __ms
-- | -- | -- | -- | --
2446 | 1969 | 84 | 8 | --
2467 | 2056 | 70 | 8 | --
2541 | 2060 | 104 | 12 | --
2492 | 2084 | 77 | 8 | --
2797 | 2369 | 74 | 9 | --

### AFTER: Load times to see the map w/ circles

total | scripts (all the yellow) | rendering (all the purple) | painting (all the green) | 'load' handler took __ms
-- | -- | -- | -- | --
1037 | 899 | 46 | 5 | --
1129 | 882 | 36 | 5 | --
1217 | 950 | 41 | 8 | --
1181 | 928 | 41 | 5 | --
1200 | 925 | 40 | 6 | --

## Times for first transition from Total to Pie view
As expected, these didn't really change much since the sorting was happening within the regular circles and I don't think that affected pies.

### BEFORE

total | scripts (all the yellow) | rendering (all the purple) | painting (all the green) | 'mouseover' handler
-- | -- | -- | -- | --
-- | 291 | 216 | 23 | 291
-- | 366 | 200 | 24 | 294
-- | 402 | 228 | 26 | 258
-- | 358 | 218 | 26 | 308
-- | 357 | 196 | 24 | 279

### AFTER

total | scripts (all the yellow) | rendering (all the purple) | painting (all the green) | 'mouseover' handler
-- | -- | -- | -- | --
-- | 364 | 195 | 25 | 332
-- | 319 | 200 | 27 | 260
-- | 377 | 200 | 28 | 317
-- | 476 | 210 | 27 | 442
-- | 354 | 197 | 26 | 321

## Times for subsequent transitions from Total to Pie view
As expected, these didn't really change much since the sorting was happening within the regular circles and I don't think that affected pies.

### BEFORE

total | scripts (all the yellow) | rendering (all the purple) | painting (all the green) | 'mouseover' handler
-- | -- | -- | -- | --
-- | 211 | 165 | 23 | 129
-- | 131 | 169 | 26 | 131
-- | 139 | 171 | 24 | 139
-- | 161 | 179 | 27 | 161
-- | 136 | 166 | 22 | 136

### AFTER

total | scripts (all the yellow) | rendering (all the purple) | painting (all the green) | 'mouseover' handler
-- | -- | -- | -- | --
-- | 149 | 169 | 24 | 149
-- | 136 | 166 | 25 | 134
-- | 144 | 177 | 33 | 144
-- | 134 | 178 | 26 | 131
-- | 148 | 199 | 30 | 146

## Times for transitions among circle categories (mouseovers)

### BEFORE

total | scripts (yellow) | rendering (purple) | painting (green) | 'mouseover' handler | from what to what
-- | -- | -- | -- | -- | --
-- | 126 | 24 | 4 | 112 | from blue to blue!
-- | 57 | 34 | 2 | 56 | blue to yellow
-- | 88 | 42 | 6 | 86 | blue to red
-- | 73 | 30 | 5 | 72 | blue to green
-- | 60 | 35 | 4 | 60 | blue to grey

### AFTER

total | scripts (yellow) | rendering (purple) | painting (green) | 'mouseover' handler | from what to what
-- | -- | -- | -- | -- | --
-- | 23 | 20 | 6 | 22 | from blue to blue!
-- | 25 | 29 | 3 | 23 | blue to yellow
-- | 27 | 35 | 4 | 25 | blue to red
-- | 25 | 33 | 5 | 24 | blue to green
-- | 24 | 30 | 3 | 24 | blue to grey

## Times for transitions among circle categories (mouseouts)

### BEFORE

total | scripts (yellow) | rendering (purple) | painting (green) | 'mouseout' handler | from what to what
-- | -- | -- | -- | -- | --
-- | 70 | 26 | 4 | 70 | blue to blue
-- | 102 | 40 | 5 | 102 | yellow to blue
-- | 107 | 33 | 4 | 100 | red to blue
-- | 86 | 34 | 4 | 85 | green to blue
-- | 87 | 32 | 4 | 86 | grey to blue

### AFTER

total | scripts (yellow) | rendering (purple) | painting (green) | 'mouseout' handler | from what to what
-- | -- | -- | -- | -- | --
-- | 15 | 21 | 5 | 14 | blue to blue
-- | 23 | 51 | 24 | 22 | yellow to blue
-- | 17 | 33 | 4 | 16 | red to blue
-- | 23 | 30 | 4 | 23 | green to blue
-- | 23 | 47 | 4 | 22 | grey to blue