@lindsaycarr 2018-03-30
Also, this touched #78 (not exactly the same fix that this requested, but tooltips added to counties and removed from points), #97 (category switching excluding pies are much faster), and #92 (not exactly the desired behavior listed here, but mousing over a county and clicking it will zoom you to the state).