@lindsaycarr 2018-03-30
I agree that the diagnostic tables could be presented in a different way. For sake of comparison, I continued with what was added in #97. I know these have a lot more information than they probably need, but that's mostly because I'm still learning which of these things we can affect and which we can't.

I'm 98% sure that transition improvement is due to no sorting and re-sorting between categories. That sorting function took up a lot of time when we were looking at performance yesterday which is what made us decide to try it out. 

The only other thing that changed for the circles was removing their mouseover and mouseout attributes. This might of had a small effect on the initial circle load, because it was only being added in the initial circle enter. But since we didn't _remove_ the mouse event, just moved them to a different element (the county boundaries), I don't think that actually saved any time.